created prefabs, can undo building the level, tilesets are now not spawned as prefabs, added some new properties for layer definition, bug fix when injecting entities from a null value. Data storage method for tiles are on the way

+++ Assets/LDtkUnity/Samples~/Entities/Scripts/ExamplePlayer.cs
@@ -5,7 +5,7 @@ namespace Samples.Entities
 {
     public class ExamplePlayer : MonoBehaviour
     {
-        [LDtkField("life")] public float _life;
+        [LDtkField("life")] public int _life;
         [LDtkField("isAwaken")] public bool _isAwaken;
         [LDtkField("weapon")] public ItemType _weapon;
         [LDtkField("bag")] public ItemType[] _bag;
